Use official GitHub Node.gitignore
https://github.com/github/gitignore/blob/main/Node.gitignore

Using #region and #endregion to delimit it as both a comment and so
other #regions can be added if need to combine Node.gitignore with
something like Python.gitignore. Can also add custom things at the top
outside of the #region stuff.

The #region thing is really just a convention. @jcbhmr uses it a lot.