Add typescript & vite to package.json
I am following a mix of:
- https://github.com/jasonsturges/vite-typescript-npm-package
- https://onderonur.netlify.app/blog/creating-a-typescript-library-with-vite/

I'm not sure how well this will turn out, but we'll see if I can get
Vite to work as a TypeScript build tool... 😉

This my first time with Vite. Why Vite? It's the top of the JS 2022
survey, so there's no time like now to jump on the Vite hype train. I'm
learning while doing. 🧠