Revamp readme
I'm using a bunch of badges here. The installation badges are supposed
to indicate what platforms this thing is available on to download from.
The usage badges are intended to indicate what platforms/things this package
can be used with. The development icons are intended to highlight developer
technologies that this package uses. This is all sort of my convention for readmes:
use image badges instead of text where possible, and use example images/code
where applicable. A picture is worth 1000 words after all!

For the <table> formatting, I didn't really know what to do. Prettier doesn't really seem to be
much help in this instance, so I just tried to squish all the table stuff onto one
line in between the actual Markdown-ish code.