Move to Vite (which still uses Rollup)
⚠️ I haven't been able to confirm that this passes the test.js suite yet!
I can't run the npx headless-test command on my Codespaces instance 😢.

This should produce the same tarball bundle as the rollup version, just with ts
sprinkled in with a different build tool.

This version also uses terser instead of the google closure compiler. Hopefully
this is OK and doesn't need to be reverted.

I also switched up the npm scripts so that the build is hooked into the packing operation.
You can restore the original build behaviour by adding an alias to it (or aliasing the
prepack to npm run build, take your pick).

I also switched to modern exports package.json syntax instead of the main entry.
This is somewhat a stylistic choice, but it also seems to be the way the Node.js
world is headed? At least it makes the module key obsolete![^1]

[^1]: https://stackoverflow.com/a/75097813