Add basic playwright test
This took some Googling and such, but I finally got the playwright test command to work.
It successfully ignores the OLD-test.js (which it apparently didn't when it was in the root as test.js)
and it also spins up an actual headless (for now at least) browser that runs the HTML file
from a basic dev server (hence the serve dep).

In the future, there should be more .test.js or .mjs files to add more coverage from the original
OLD-test.js file back into the testing loop.