Add GitHub Action to test main branch
This will continuously test the main branch and provide a GREEN or RED (or YELLOW) symbol
next to each commit that has this check pass/fail on it. This will also add a similar
thing to all PRs that target the main branch.

The npm commands are npm ci && npm test. This COULD be shortened to npm cit which is
a shortcut command to run npm ci && npm test, but I think that this more verbose version is better
for readability.

⚠ I'm not sure if the GitHub Actions runners have the needed Chome/whatever installed to
actually run a headless browser?