Move to a more basic WPT-like setup
The way web platform tests usually work is with an HTML page that runs some stuff and reports
things IN THE BROWSER with access to all the fancy browser-level APIs like dom, fetch,
XMLSerializer, etc. Since this is sorta-ish like a WPT use-case (testing something on
the web platform, not testing existing UI) it makes sense to adopt something
similar. This wpt-like.mjs script is very bare and doesn't do a whole lot. It just
makes sure that there are no errors on any of the HTML pages after they have run.