Use test/ instead of demo/ for GitHub Pages
Maybe it would be a good idea to use something like CodeSandbox or CodePen to host
a demo site so that users can quickly play around and edit it? I would think that might be more
interactive and engaging, but a static demo site also works.

I removed the original demo files since I tried to re-create much of the headless
testing (that used the demo files) in the test/ folder with the wpt-like test runner.
This means that all of those things can be run in a plain browser and the user
can see the demo thing working-ish.

This is probably the most controversial change that I've made so far.