Add stub contributing guide
Based on my current generic contributing.md file that I use. This is quite
a dynamic document that can go in many different directions.
https://github.com/jcbhmr/.github/blob/main/CONTRIBUTING.md

My jcbhmr/.github contributing.md is originally based on the https://generator.contributing.md/ but
with massive cuts to the text to make it shorter. In the future this can/should be expanded
when/if the process to add new stuff to this repo changes.

https://mozillascience.github.io/working-open-workshop/contributing/