Add edge cases docs and spec compliance docs
The Edge cases docs are intended for devs who want to know what cases are covered
and how they are covered. It is probably the lesser of these two documents in terms of importance, but
it does have some quality information about how certain potential issues might
manifest in this project.

The spec compliance document is more for referencing where the standard that this project
is based on comes from. It's intended for devs who want to add code/change code and
are wondering what resources there are about the actual implementation of
<style scoped>.