Add my own how it works for devs
This is intended to be a document describing how the project works and what parts are
where and how they fit together. This isn't really a user facing document. It's
more for devs adding/changing/removing code.

For user docs about how this works, I'd put that in the readme.