Remove options
Moving forward, I don't think these options really make sense.

I'd love some rationalization as to why they're important, but for my rewrite I'm
not going to focus on them. Hence the removal from docs.

This is part of the push to rewrite and improve the code. I'd like to get v1.0.0
or something off the ground!